Se corrigio el reporte tiempo de uso

+++ src/main/java/com/velociteam/pspecs/dto/ReportDTO.java
@@ -7,20 +7,20 @@ import com.velociteam.pspecs.services.Tuple;
 
 public class ReportDTO {
 	
-	Map<String,Long> tiemposDeUso;
+	Map<String,Double> tiemposDeUso;
 	Map<String,List<Tuple>> usuariosContactados;
 	Map<String,List<Tuple>> pictogramasMasUtilizados;
 
 	public ReportDTO() {
 	}
 	
-	public ReportDTO(Map<String, Long> tiemposDeUso, Map<String, List<Tuple>> usuariosContactados,Map<String,List<Tuple>> pictogramasMasUtilizados) {
+	public ReportDTO(Map<String, Double> tiemposDeUso, Map<String, List<Tuple>> usuariosContactados,Map<String,List<Tuple>> pictogramasMasUtilizados) {
 		this.tiemposDeUso = tiemposDeUso;
 		this.usuariosContactados = usuariosContactados;
 		this.pictogramasMasUtilizados = pictogramasMasUtilizados;
 	}
 	
-	public Map<String, Long> getTiemposDeUso() {
+	public Map<String, Double> getTiemposDeUso() {
 		return tiemposDeUso;
 	}
